requirements.txt
# Core
requests==2.31.0
pandas==2.1.4
python-dotenv==1.0.0

# Database
SQLAlchemy==2.0.25

# Orchestration
apache-airflow==2.8.1

# Testing
pytest==7.4.4

# Linting
flake8==7.0.0
